requirements.txt
numpy
opencv-python
scipy
matplotlib
pyyaml
